refactor(libs): add typedoc comment to libs func
