+++ package.json
@@ -1,9 +1,9 @@
 {
     "name": "lemon-typescript-template",
-    "version": "1.0.2",
+    "version": "1.0.4",
     "description": "lemon npm module template",
-    "main": "dist/app.js",
-    "types": "dist/app.d.ts",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
     "scripts": {
         "nodemon": "NODE_ENV=development nodemon dist/app.js",
         "start": "npm run build && concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run nodemon\"",
