+++ src/controllers/hello/index.ts
@@ -4,7 +4,7 @@ const router: Router = Router();
 
 router.get('/', (_: Request, res: Response) => {
     res.json({
-        result: 'Hello Typescript!!',
+        result: 'Hello Lemon',
     });
 });
