--- src/libs/index.ts
@@ -1,2 +1,0 @@
-export const sayHello = (name: string) => `Hello ${name}`;
-export const sum = (list: number[]) => list.reduce((acc, val) => acc + val);
